requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
gunicorn==20.0.4
huey==2.3.0
idna==2.10
importlib-metadata==3.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
phonenumbers==8.12.18
prompt-toolkit==3.0.16
PyJWT==1.7.1
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
twilio==6.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
